+++ src/zino/trapd/__init__.py
@@ -1,0 +1,3 @@
+"""Zino SNMP trap daemon back-ends"""
+
+from zino.trapd.pysnmp_backend import *  # noqa
